requirements.txt
aiohttp==3.8.4
aioinflux==0.9.0
alpaca_trade_api==3.0.2
apache-airflow==2.8.0
arrow==1.2.2
binance==0.3
boto3==1.24.28
botocore==1.27.28
chromedriver_autoinstaller==0.6.3
click==8.0.4
cryptocompare==0.7.6
cvxpy==1.4.1
dacite==1.8.1
deprecated==1.2.14
humanize==4.9.0
ib_insync==0.9.86
ipython==7.31.1
loguru==0.7.2
matplotlib==3.5.2
mode==4.4.0
numpy==1.21.5
opsgenie_sdk==2.1.5
overrides==7.4.0
pandas==1.4.4
pandas_datareader==0.10.0
pandas_market_calendars==4.2.1
plotly==5.9.0
psutil==5.9.0
pyarrow==12.0.1
pytest==7.1.2
python-dotenv==1.0.0
python_dateutil==2.8.2
pytz==2023.3
rich==13.7.0
scipy==1.9.1
seaborn==0.11.2
selenium==4.16.0
statsmodels==0.13.2
tabulate==0.8.10
tenacity==8.0.1
tqdm==4.64.1
